+++ Expanded Cheat Menu/Code/Menus/ExpandedMenu.lua
@@ -8,7 +8,7 @@ function OnMsg.ClassesGenerate()
 
 	c = c + 1
 	Actions[c] = {ActionName = S[302535920000031--[[Find Nearest Resource--]]],
-		ActionMenubar = "ECM.Expanded CM",
+		ActionMenubar = "ECM.ECM",
 		ActionId = ".Find Nearest Resource",
 		ActionIcon = "CommonAssets/UI/Menu/EV_OpenFirst.tga",
 		RolloverText = S[302535920000554--[[Select an object and click this to display a list of resources.--]]],
@@ -20,7 +20,7 @@ function OnMsg.ClassesGenerate()
 
 	c = c + 1
 	Actions[c] = {ActionName = S[302535920000333--[[Building Info--]]],
-		ActionMenubar = "ECM.Expanded CM",
+		ActionMenubar = "ECM.ECM",
 		ActionId = ".Building Info",
 		ActionIcon = "CommonAssets/UI/Menu/ExportImageSequence.tga",
 		RolloverText = S[302535920000345--[[Shows info about building in text above it.--]]],
@@ -30,7 +30,7 @@ function OnMsg.ClassesGenerate()
 
 	--~ c = c + 1
 	--~ Actions[c] = {ActionName = S[302535920000555--[[Monitor Info--]]],
-	--~ 	ActionMenubar = "ECM.Expanded CM",
+	--~ 	ActionMenubar = "ECM.ECM",
 	--~ 	ActionId = ".Monitor Info",
 	--~ 	ActionIcon = "CommonAssets/UI/Menu/EV_OpenFirst.tga",
 	--~ 	RolloverText = S[302535920000556--[[Shows a list of updated information about your city.--]]],
@@ -40,7 +40,7 @@ function OnMsg.ClassesGenerate()
 
 	c = c + 1
 	Actions[c] = {ActionName = S[302535920000469--[[Close Dialogs--]]],
-		ActionMenubar = "ECM.Expanded CM",
+		ActionMenubar = "ECM.ECM",
 		ActionId = ".Close Dialogs",
 		ActionIcon = "CommonAssets/UI/Menu/remove_water.tga",
 		RolloverText = S[302535920000470--[[Close any dialogs opened by ECM (Examine, Object Editor, Change Colours, etc...)--]]],